requirements.txt
pandas==2.3.3
geopandas==1.1.3
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
folium==0.20.0
shapely==2.1.2
requests==2.32.5
jupyter==1.1.1
openpyxl==3.1.5
pyproj==3.7.2
python-docx==1.2.0
